fix(viz): numeric idiom lints in median/parity and area-series code
Batch J1 of resinsim-viz-clippy-baseline. Six errors, judgment-tier.

- manual_is_multiple_of (3x): the even-length median branches in
  cure_depth.rs::median_finite and forces.rs::median become
  len().is_multiple_of(2); the trailing-zero strip loop in
  scrubber.rs's tick_step_returns_round_numbers test becomes
  s.is_multiple_of(10), && s > 9 untouched. All three operands are
  unsigned (usize len(), u32 step), so is_multiple_of is defined and
  the rewrite is a pure predicate-shape swap — the median parity
  branches still select the same element on odd and even lengths.

- useless_conversion (2x): area_delta.rs's draw_loaded wrapped
  layer.cross_section_area_mm2 and layer.area_delta_mm2 (both
  LayerResult fields, already f64) in f64::from(). Conversion dropped
  at both sites; a short comment now names the f64 assumption at each
  site so a future narrowing to f32 fails loudly instead of silently
  changing meaning with no lint to catch it. Subsequent is_finite
  guards and the plotted-point construction are unchanged.

- manual_range_contains: scrubber.rs's tick_step_lilith_scale test
  assert!(step >= 500 && step <= 1000, ...) becomes
  assert!((500..=1000).contains(&step), ...), same message.

Remaining resinsim-viz clippy errors: 26 (cargo clippy -p resinsim-viz
--all-targets -- -D warnings).

+++ crates/resinsim-viz/src/ui/v2/panes/area_delta.rs
@@ -35,11 +35,13 @@ fn draw_loaded(ui: &mut egui::Ui, sim: &PrintSimulation, ctx: &PaneCtx<'_>) {
     let mut delta: Vec<[f64; 2]> = Vec::with_capacity(n);
     for (i, layer) in sim.layers().iter().enumerate() {
         let x = i as f64;
-        let a = f64::from(layer.cross_section_area_mm2);
+        // cross_section_area_mm2 is already f64 (LayerResult); no widening needed.
+        let a = layer.cross_section_area_mm2;
         if a.is_finite() {
             area.push([x, a]);
         }
-        let d = f64::from(layer.area_delta_mm2);
+        // area_delta_mm2 is already f64 (LayerResult); no widening needed.
+        let d = layer.area_delta_mm2;
         if d.is_finite() {
             delta.push([x, d]);
         }
